API

loading in data

Peccon.data_alpha — Function

data_alpha(Tickers)

extracts the daily price info of multiple stocks from alphavantage and puts them in a vector of dataframes.

Examples

julia> data_alpha(["ADAEUR", "SPY"])

General

Calculating returns

Peccon.daily_returns — Method

daily_returns(portfolio, Tickers)

calculates the daily log returns of each stock in a portfolio based on the close price of the day.

Examples

julia> tickers = ["ADAEUR", "SPY"]
julia> data = fin_data(tickers)
julia> daily_returns(data, tickers)

Peccon.per_return — Method

per_return(returns)

calculates the compounded return for a specific time-period from daily log returns

# Examples

julia> tickers = ["ADAEUR", "SPY"]
julia> data = fin_data(tickers)
julia> calc_returns(data, tickers)
julia> data_alpha(["ADAEUR", "SPY"])

Tools

modern portfolio theory (mpt)

Peccon.sharp_ratio — Function

sharp_ratio(port_sim)

calculates the sharp ratio of each simulates portfolio

Examples

julia> port_sim = sim_mpt(stock_returns)
julia> sharp_ratio(port_sim) 

Peccon.sim_mpt — Function

sim_opt(returns, simulations= 5000, days=252)

simulates random portfolio combinations and calculates the expected return and standard deviation of the portfolio

Examples

julia> returns = daily_returns(data, tickers)
julia> sim_mpt(returns)

Peccon.opt_mpt — Function

opt_mpt(returns, risk_av_step = 0.0:0.02:2.0, diversification_limit= 0.05)

returns the efficient frontier for a portfolio.

Examples

julia> port_opt = opt_mpt(returns)